backend/requirements.txt
pymongo
fastapi
fastapi-mail
python-decouple
bcrypt
cryptography
passlib
pydantic
pyjwt
pytz
uvicorn
httpx
